Move radio service code verbatim into `repository/radio`
