+++ app/src/main/AndroidManifest.xml
@@ -117,7 +117,7 @@
 
         <!-- This is a private service which just does direct communication to the radio -->
         <service
-            android:name="com.geeksville.mesh.service.RadioInterfaceService"
+            android:name="com.geeksville.mesh.repository.radio.RadioInterfaceService"
             android:enabled="true"
             android:exported="false" />
