+++ app/src/main/java/com/geeksville/mesh/repository/radio/IRadioInterface.kt
@@ -1,4 +1,4 @@
-package com.geeksville.mesh.service
+package com.geeksville.mesh.repository.radio
 
 import java.io.Closeable
