+++ app/src/main/java/com/geeksville/mesh/repository/radio/RadioInterfaceService.kt
@@ -1,4 +1,4 @@
-package com.geeksville.mesh.service
+package com.geeksville.mesh.repository.radio
 
 import android.annotation.SuppressLint
 import android.app.Service
@@ -17,6 +17,10 @@ import com.geeksville.concurrent.handledLaunch
 import com.geeksville.mesh.IRadioInterfaceService
 import com.geeksville.mesh.repository.bluetooth.BluetoothRepository
 import com.geeksville.mesh.repository.usb.UsbRepository
+import com.geeksville.mesh.service.EXTRA_CONNECTED
+import com.geeksville.mesh.service.EXTRA_PAYLOAD
+import com.geeksville.mesh.service.EXTRA_PERMANENT
+import com.geeksville.mesh.service.prefix
 import com.geeksville.util.anonymize
 import com.geeksville.util.ignoreException
 import com.geeksville.util.toRemoteExceptions
@@ -122,7 +126,8 @@ class RadioInterfaceService : Service(), Logging {
             if (address != null) {
                 val c = address[0]
                 val rest = address.substring(1)
-                val isValid = InterfaceFactory.getFactory(c)?.addressValid(context, usbRepository, rest) ?: false
+                val isValid = InterfaceFactory.getFactory(c)
+                    ?.addressValid(context, usbRepository, rest) ?: false
                 if (!isValid)
                     return null
             }
