+++ app/src/main/java/com/geeksville/mesh/repository/radio/SerialInterface.kt
@@ -1,4 +1,4 @@
-package com.geeksville.mesh.service
+package com.geeksville.mesh.repository.radio
 
 import android.content.Context
 import com.geeksville.android.Logging
